third_party/blink/web_tests/platform/generic/fast/media/mq-display-mode-fullscreen-expected.txt

Test the (display-mode: ...) media features while the browser window entering/leaving fullscreen mode.

------------- before entering fullscreen -------------
Query "(display-mode)": true
Query "(display-mode: browser)": true
Query "(display-mode: minimal-ui)": false
Query "(display-mode: standalone)": false
Query "(display-mode: fullscreen)": false
------------- entered fullscreen -------------
Query "(display-mode)": true
Query "(display-mode: browser)": false
Query "(display-mode: minimal-ui)": false
Query "(display-mode: standalone)": false
Query "(display-mode: fullscreen)": true
------------- left fullscreen -------------
Query "(display-mode)": true
Query "(display-mode: browser)": true
Query "(display-mode: minimal-ui)": false
Query "(display-mode: standalone)": false
Query "(display-mode: fullscreen)": false
